backend/requirements.txt
asgiref==3.9.1
psycopg2-binary==2.9.10
asyncpg==0.30.0
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
django-celery-beat==2.8.0
celery==5.4.0
redis==5.2.1